requirements.txt
# Deep Learning & PyTorch
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# MLOps
mlflow>=2.8.0

# API Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Interface Streamlit
streamlit>=1.28.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0

# Utilitaires
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
